requirements.txt
-e ./fairseq/
opencv-python
timm
ftfy==6.0.3
tensorboardX==2.4.1
pycocotools
pycocoevalcap==1.2
pytorch_lightning
einops
datasets
rouge_score
soundfile
editdistance
librosa
python-Levenshtein
zhconv
pypinyin==0.47.1
g2p_en
torchvision
transformers
tqdm
pandas
regex
braceexpand
webdataset
jsonlines
